Clean up experiments and unused code for MVP
