+++ packages/edit-site/src/components/style-book/test/utils.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { getCategoryExamples, getExamples, getBlockCategories } from '../utils';
+import { getCategoryExamples, getExamples } from '../utils';
 import { STYLE_BOOK_CATEGORIES } from '../constants';
 
 /**
@@ -9,55 +9,6 @@ import { STYLE_BOOK_CATEGORIES } from '../constants';
  */
 import { getBlockTypes, unregisterBlockType } from '@wordpress/blocks';
 import { registerCoreBlocks } from '@wordpress/block-library';
-
-// Fixtures
-/*const exampleBlocks = [
-	{
-		name: 'core/post-content',
-		title: 'Post Content',
-		category: 'theme',
-	},
-	{
-		name: 'core/home-link',
-		title: 'Home Link',
-		category: 'design',
-	},
-	{
-		name: 'custom/colors',
-		title: 'Colors',
-		category: 'colors',
-	},
-	{
-		name: 'core/site-logo',
-		title: 'Site Logo',
-		category: 'theme',
-	},
-	{
-		name: 'core/site-title',
-		title: 'Site Title',
-		category: 'theme',
-	},
-	{
-		name: 'core/site-tagline',
-		title: 'Site Tagline',
-		category: 'theme',
-	},
-	{
-		name: 'core/group',
-		title: 'Group',
-		category: 'design',
-	},
-	{
-		name: 'core/comments-pagination-numbers',
-		title: 'Comments Page Numbers',
-		category: 'theme',
-	},
-	{
-		name: 'core/post-featured-image',
-		title: 'Featured Image',
-		category: 'theme',
-	},
-];*/
 
 describe( 'utils', () => {
 	beforeAll( () => {
@@ -81,18 +32,10 @@ describe( 'utils', () => {
 				themeCategory,
 				getExamples()
 			);
-			expect(
-				themeCategoryExamples[ 'site-identity' ].examples
-			).toBeDefined();
-			expect( themeCategoryExamples.design.examples ).toBeDefined();
-			expect( themeCategoryExamples.posts.examples ).toBeDefined();
-			expect( themeCategoryExamples.comments.examples ).toBeDefined();
+			expect( themeCategoryExamples.name ).toEqual( 'theme' );
+			expect( themeCategoryExamples.subcategories ).toHaveLength(
+				themeCategory.subcategories.length
+			);
 		} );
 	} );
-
-/*	describe( 'getBlockCategories', () => {
-		it( 'returns categories', () => {
-			expect( getBlockCategories( getExamples() ) ).toEqual( ' ' );
-		} );
-	} );*/
 } );
