+++ packages/edit-site/src/components/style-book/utils.js
@@ -1,101 +1,13 @@
 /**
- * External dependencies
- */
-
-/**
  * WordPress dependencies
  */
 import { __, sprintf } from '@wordpress/i18n';
 import {
-	getCategories,
 	getBlockType,
 	getBlockTypes,
 	getBlockFromExample,
 	createBlock,
 } from '@wordpress/blocks';
-import { __experimentalGetCoreBlocks } from '@wordpress/block-library';
-import { privateApis as editorPrivateApis } from '@wordpress/editor';
-import { useMemo, useState, memo, useContext } from '@wordpress/element';
-
-/**
- * Internal dependencies
- */
-import { unlock } from '../../lock-unlock';
-import {
-	STYLE_BOOK_CATEGORIES,
-	STYLE_BOOK_THEME_SUBCATEGORIES,
-} from './constants';
-import { privateApis as blockEditorPrivateApis } from '@wordpress/block-editor';
-
-const {
-	ExperimentalBlockEditorProvider,
-	useGlobalStyle,
-	GlobalStylesContext,
-	useGlobalStylesOutputWithConfig,
-} = unlock( blockEditorPrivateApis );
-
-/*
-	// This hook will get all block examples,
-	// and group them by category.
-	- get all registered block examples. See current getExamples()
-	- build any adhoc examples.
-	STYLE_BOOK_CATEGORIES
-
- */
-
-function getCustomExamples() {
-	const customExamples = [
-		{
-			name: 'custom/colors',
-			title: __( 'Colors' ),
-			category: 'colors',
-			blocks: [
-				createBlock( 'core/group', {
-					attributes: {
-						layout: {
-							type: 'grid',
-							columnCount: 3,
-						},
-					},
-					innerBlocks: [
-						createBlock( 'core/group', {
-							style: {
-								layout: {
-									selfStretch: 'fill',
-								},
-							},
-							layout: {
-								type: 'constrained',
-							},
-						} ),
-					],
-				} ),
-			],
-		},
-	];
-}
-
-// @TODO get other registered categories (aside from what's in STYLE_BOOK_CATEGORIES)
-// And their subcategories.
-export function getBlockCategories( examples ) {
-	const reservedCategories = [
-		...STYLE_BOOK_THEME_SUBCATEGORIES,
-		...STYLE_BOOK_CATEGORIES,
-	].map( ( category ) => category.category );
-
-	return getCategories()
-		.filter(
-			( category ) =>
-				! reservedCategories.includes( category.slug ) &&
-				examples.some(
-					( example ) => example.category === category.slug
-				)
-		)
-		.map( ( category ) => ( {
-			name: category.slug,
-			title: category.title,
-		} ) );
-}
 
 // Get all examples from every registered block.
 export function getExamples() {
@@ -114,31 +26,35 @@ export function getExamples() {
 			category: blockType.category,
 			blocks: getBlockFromExample( blockType.name, blockType.example ),
 		} ) );
-	/*
-	 * Use our own example for the Heading block so that we can show multiple
-	 * heading levels.
-	 */
-	const headingsExample = !! getBlockType( 'core/heading' )
-		? {
-				name: 'core/heading',
-				title: __( 'Headings' ),
-				category: 'text',
-				blocks: [ 1, 2, 3, 4, 5, 6 ].map( ( level ) => {
-					return createBlock( 'core/heading', {
-						content: sprintf(
-							// translators: %d: heading level e.g: "1", "2", "3"
-							__( 'Heading %d' ),
-							level
-						),
-						level,
-					} );
-				} ),
-		  }
-		: {};
+
+	const isHeadingBlockRegistered = !! getBlockType( 'core/heading' );
+
+	if ( ! isHeadingBlockRegistered ) {
+		return nonHeadingBlockExamples;
+	}
+
+	// Use our own example for the Heading block so that we can show multiple
+	// heading levels.
+	const headingsExample = {
+		name: 'core/heading',
+		title: __( 'Headings' ),
+		category: 'text',
+		blocks: [ 1, 2, 3, 4, 5, 6 ].map( ( level ) => {
+			return createBlock( 'core/heading', {
+				content: sprintf(
+					// translators: %d: heading level e.g: "1", "2", "3"
+					__( 'Heading %d' ),
+					level
+				),
+				level,
+			} );
+		} ),
+	};
 
 	return [ headingsExample, ...nonHeadingBlockExamples ];
 }
 
+// Get examples for a given category.
 export function getCategoryExamples( categoryDefinition, examples ) {
 	if ( ! categoryDefinition?.name || ! examples?.length ) {
 		return [];
@@ -163,9 +79,6 @@ export function getCategoryExamples( categoryDefinition, examples ) {
 	const blocksToInclude = categoryDefinition?.blocks || [];
 	const blocksToExclude = categoryDefinition?.exclude || [];
 
-	// @TODO maybe return an array. [ { label: string, examples: [] } ]
-	// With subcategories { subcategories: [ { title: string, examples: [] } ] }
-	// With single category examples { title: string,  examples: [] }
 	return {
 		title: categoryDefinition.title,
 		name: categoryDefinition.name,
